+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Improved skin names for most models to fit on the screen.
 - Improved default selected geosets for models.
 - Fixed missing escape slashes in regex guide tooltip.
+- Added a 3D grid to the model viewer (can be toggled on the sidebar).
 
 0.1.23 (28-11-2020)
 - Added a warning to the configuration screen for export directories that contain spaces.